feat(desktop): move sort chip options into a popover
